svn merge -r 1966:1976 svn://svn.minisip.org/var/svn/minisip/branches/ppc_evc4_cesc/minisip .
minisip compiles in MS EVC++ 4.0 !!!

TODO: 
- Create a nice gui ... this way we can test the functionality of the code (now it compiles .. .that's it :D )
	- options ... directly in EVC 4.0 ...
	- .NET ... this way it could be used in non windows ce platforms ... (not crosscompiling with mingw32, though).
- libminisip ... or not
- documentation over the port and instructions
- add the base external dependencies ... includes, web links to src code, patches needed, compiled libraries (if possible).




git-svn-id: svn://svn.minisip.org/minisip/trunk@1981 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ minisip/aec/aec.h
@@ -73,7 +73,7 @@ class IIR6 {
 public:
   IIR6();
   float highpass(float in) {
-    const float AlphaHp = 0.075;  /* controls Transfer Frequence */
+    const float AlphaHp = (float)0.075;  /* controls Transfer Frequence */
     const float Gain6   = 1.45f;  /* gain to undo filter attenuation */
 
     /* Highpass = Signal - Lowpass. Lowpass = Exponential Smoothing */
